Fix some style issues in Puppet manifests
These checks are performed on the Puppet Forge to give quality scores to
modules, and are failing for the choria repositories:
* The `ensure` parameter must be the first one in resource declaration;
* Non-interpolated strings should be single, not double-quoted.

A bunch of other issues are reported by puppet-lint, but as of today
they do not seem to be part of the module score calculation, so in order
to minimize changes, this commit do not address them.

While these changes are gratuitous, they will allow the scoring of the
modules to increase, giving users a more positive view of the care taken
to maintain them.